requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==4.1.2
idna==3.4
pycparser==2.21
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
django-auditlog==2.2.2
django-easy-audit==1.3.3
python-dotenv
django-admin-geomap
selenium